% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterCCC}
\alias{filterCCC}
\title{Filter the CCC results}
\usage{
filterCCC(
  data,
  ligand.filter = "logFC > 0.3 & padj < 0.05",
  receptor.filter = NULL
)
}
\arguments{
\item{data}{An object returned by the multiCCC or filterCCC function.

Filtering is always applied to the original result table of the input object, preventing cumulative filtering effects.}

\item{ligand.filter}{Provide a string with the rule defined by the subset function to filter CCC based on the expression of ligands. If NULL, no filtering will be performed.

Available filtering metrics include feature, group, avgExpr, logFC, statistic, auc, pval, padj, pct_in, and pct_out.}

\item{receptor.filter}{Similar to the ligand.filter parameter, provide a string with the rule defined by the subset function to filter CCC based on the expression of receptors. If NULL, no filtering will be performed.}
}
\value{
The main output content is stored in 'gene.expression' and 'filtered.result'. 'gene.expression' represents the evaluation results of cell-specific markers, and 'filtered.result' represents the filtered CCC results.

Note: The filterCCC function simultaneously applies filtering to all analysis results, without requiring separate filtering for different application scenarios.
}
\description{
Perform the Wilcoxon rank sum test to identify cell-specific markers, and then use the resulting gene list for filtering the CCC results.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
#Note: The multiCCC function supports the simultaneous use of the binary.params, anova.column, glm.column, and time.course.params parameters, and analyses will be performed automatically for each application scenario.
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )
ccc.glm <- multiCCC( data = LRscore , glm.column = 'weight' )
ccc.time <- multiCCC( data = LRscore , time.course.params = list( time  = 'time' , replicate  = 'replicate'  ) )

#One-step execution for all scenarios
ccc.all <- multiCCC( data = LRscore, 
	   binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ),
	   anova.column = 'batch',
	   glm.column = 'weight',
	   time.course.params = list( time  = 'time' , replicate  = 'replicate'  )
	)

#3.filterCCC
#Note: The filterCCC function simultaneously applies filtering to all analysis results, without requiring separate filtering for different application scenarios.
filter.binary <- filterCCC(ccc.binary)
filter.anova <- filterCCC(ccc.anova)
filter.glm <- filterCCC(ccc.glm )
filter.time <- filterCCC(ccc.time)

#One-step execution for all scenarios
filter.all <- filterCCC(ccc.all)

}
