% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_score.R
\name{lr_score}
\alias{lr_score}
\title{Compute cell–cell communication score (LRscore)}
\usage{
lr_score(
  exp,
  meta.data,
  sample,
  celltype,
  LR.species = "human",
  LR.source = "CCI",
  LR.method = "CCI",
  min.cell = 10,
  min.exp = 0.1,
  min.prob = 0.3,
  threads = NULL
)
}
\arguments{
\item{exp}{A matrix object with cells as row names and genes as column names.}

\item{meta.data}{A data.frame object whose row names are identical to those of the exp parameter.}

\item{sample}{The column name in meta.data that represents the sample ID.}

\item{celltype}{The column name in meta.data that represents the cell type.}

\item{LR.species}{Species. Options are human or mouse.}

\item{LR.source}{Ligand–receptor resource. Options are CCI or liana.}

\item{LR.method}{Method for calculating LR score. Options are CCI or SingleCellSignalR.}

\item{min.cell}{The minimum number of cells expressing the ligand or receptor.}

\item{min.exp}{The minimum expression level of ligands and receptors included in the analysis.}

\item{min.prob}{The minimum proportion of cells expressing the ligand and receptor.}

\item{threads}{The number of threads used for parallel computation.}
}
\value{
A list object.
}
\description{
Compute cell–cell communication score (LRscore)
}
\examples{

data( 'test.data' )
LRscore <- lr_score( exp = t( test.data$exp ) ,
                 meta.data = test.data$meta ,
                 LR.species = 'mouse' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

}
