% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.circos.R
\name{plot_circos}
\alias{plot_circos}
\title{Circos plot}
\usage{
plot_circos(
  res,
  data.used = "filtered.result",
  padj.threshold = 0.05,
  p.threshold = NULL,
  max.group = NULL,
  merge = T,
  LR = F,
  colors = NULL,
  link.alpha = 0.7,
  labels.rot = F,
  ligand = NULL,
  receptor = NULL,
  sender = NULL,
  receiver = NULL,
  CCC.ID = NULL,
  title = NULL,
  plot.title = 14
)
}
\arguments{
\item{res}{The object returned by the filterCCC function.}

\item{data.used}{The data used for plotting. Options are result or filtered.result.}

\item{padj.threshold}{The threshold for p.adj.}

\item{p.threshold}{Similar to padj.threshold, this parameter specifies the p-value threshold. When provided, padj.threshold will be ignored.}

\item{max.group}{Which group is treated as the communication-enhanced group.}

\item{merge}{Whether upregulated and downregulated CCC events are displayed in a single plot.}

\item{LR}{Whether to plot ligand–receptor pairs instead of sender–receiver pairs.}

\item{colors}{Provide a color vector to specify the fill colors of links.}

\item{link.alpha}{A numeric value specifying the transparency of the links.}

\item{labels.rot}{Whether to rotate the labels by 90 degrees.}

\item{ligand}{Character vector for filtering CCC.}

\item{receptor}{Character vector for filtering CCC.}

\item{sender}{Character vector for filtering CCC.}

\item{receiver}{Character vector for filtering CCC.}

\item{CCC.ID}{Character vector for filtering CCC.}

\item{title}{Plot title.}

\item{plot.title}{The font size of the plot title.}
}
\value{
A ggplot2 object.
}
\description{
Plot a circos plot for the results of the categorical variable.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )

#3.filterCCC
filter.binary <- filterCCC(ccc.binary)
filter.anova <- filterCCC(ccc.anova)

#4.circos
plot_circos( res = filter.binary$binary.res , p.threshold = 1   )
plot_circos( res = filter.binary$binary.res , p.threshold = 1 , merge = F  )
plot_circos( res = filter.binary$binary.res , p.threshold = 1 , LR = T  )

plot_circos( res = filter.anova$anova.res , p.threshold = 1 )

}
