% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dot}
\alias{plot_dot}
\title{Dot plot}
\usage{
plot_dot(
  res,
  data.used = "filtered.result",
  fill = "p.adj",
  threshold = 0.05,
  colors = c("#0000FF", "#F7F7F7", "#FF0000"),
  ligand = NULL,
  receptor = NULL,
  sender = NULL,
  receiver = NULL,
  CCC.ID = NULL,
  dot.max = 3,
  dot.min = 1,
  legend.position = "right",
  strip.text = 10,
  x.text = 10,
  y.text = 10,
  facet.row = 1
)
}
\arguments{
\item{res}{The object returned by the filterCCC function.}

\item{data.used}{The data used for plotting. Options are result or filtered.result.}

\item{fill}{The column used for color fill. Options are p or p.adj.}

\item{threshold}{The threshold for p or p.adj. Corresponds to the fill parameter.}

\item{colors}{A character vector.}

\item{ligand}{Character vector for filtering CCC.}

\item{receptor}{Character vector for filtering CCC.}

\item{sender}{Character vector for filtering CCC.}

\item{receiver}{Character vector for filtering CCC.}

\item{CCC.ID}{Character vector for filtering CCC.}

\item{dot.max}{The maximum value of the dots.}

\item{dot.min}{The minimum value of the dots.}

\item{legend.position}{The position of the legend in the plot.}

\item{strip.text}{The font size of the facet labels.}

\item{x.text}{The font size of the X-axis labels.}

\item{y.text}{The font size of the Y-axis labels.}

\item{facet.row}{The number of rows in the facets.}
}
\value{
A ggplot2 object.
}
\description{
Plot a dot plot for the results of the categorical variable.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )

#3.filterCCC
filter.binary <- filterCCC(ccc.binary)
filter.anova <- filterCCC(ccc.anova)

#4.plot_dot
plot_dot( res = filter.binary$binary.res ,
          ligand = c("S100A8" , "S100A8" , "CD47") ,
          threshold =  1 ,strip.text = 8 )
plot_dot( res = filter.anova$anova.res ,
          ligand =  c("HP" ,"S100A9" , "S100A8" ),
          threshold =  1 ,strip.text = 8 )
}
