% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heatmap.R
\name{plot_heatmap_cell}
\alias{plot_heatmap_cell}
\title{Heatmap of sender-receiver}
\usage{
plot_heatmap_cell(
  res,
  max.group = NULL,
  merge = T,
  trend = "Up",
  data.used = "filtered.result",
  padj.threshold = 0.05,
  p.threshold = NULL,
  ligand = NULL,
  receptor = NULL,
  sender = NULL,
  receiver = NULL,
  CCC.ID = NULL,
  top.up = NULL,
  top.down = NULL,
  top.all = NULL,
  colors = list(low = "#002BFF", mid = "#F7F7F7", high = "#FF2B00", midpoint = 0),
  border.color = "white",
  border.size = 0.5,
  x.text = 10,
  y.text = 10,
  x.title = 12,
  y.title = 12,
  title = NULL,
  plot.title = 14,
  aspect.ratio = 1,
  point.size = c(3, 6),
  point2heatmap = 0.5,
  legend.direction = "horizontal"
)
}
\arguments{
\item{res}{The object returned by the filterCCC function.}

\item{max.group}{Which group is treated as the communication-enhanced group.}

\item{merge}{Whether upregulated and downregulated CCC events are displayed in a single plot.}

\item{trend}{When merge = FALSE, selects upregulated or downregulated CCC events in max.group. Options are Up and Down.}

\item{data.used}{The data used for plotting. Options are result or filtered.result.}

\item{padj.threshold}{The threshold for p.adj.}

\item{p.threshold}{Similar to padj.threshold, this parameter specifies the p-value threshold. When provided, padj.threshold will be ignored.}

\item{ligand}{Character vector for filtering CCC.}

\item{receptor}{Character vector for filtering CCC.}

\item{sender}{Character vector for filtering CCC.}

\item{receiver}{Character vector for filtering CCC.}

\item{CCC.ID}{Character vector for filtering CCC.}

\item{top.up}{Selects the top n cell types with the highest number of upregulated CCC events for plotting.}

\item{top.down}{Selects the top n cell types with the highest number of downregulated CCC events for plotting.}

\item{top.all}{Selects the top n cell types with the most CCC events for plotting.}

\item{colors}{A list specifying the colors.}

\item{border.color}{Color of grid borders in the plot area.}

\item{border.size}{Size of grid borders in the plot area.}

\item{x.text}{The font size of the X-axis labels.}

\item{y.text}{The font size of the Y-axis labels.}

\item{x.title}{The font size of the X-axis title.}

\item{y.title}{The font size of the Y-axis title.}

\item{title}{Plot title.}

\item{plot.title}{The font size of the plot title.}

\item{aspect.ratio}{aspect ratio, expressed as y / x.}

\item{point.size}{Specifies the range of point sizes.}

\item{point2heatmap}{Ratio of the point plot to the heatmap.}

\item{legend.direction}{Legend arrangement, options: 'horizontal' or 'vertical'.}
}
\value{
A ggplot2 object.
}
\description{
Plot a heatmap for the results of the categorical variable.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )

#3.filterCCC
filter.binary <- filterCCC(ccc.binary)
filter.anova <- filterCCC(ccc.anova)

#4.heatmap
plot_heatmap_cell( res = filter.binary$binary.res , max.group = "O" , merge = T , p.threshold = 1 )
plot_heatmap_cell( res = filter.binary$binary.res , max.group = "O" , merge = F , trend = 'Up' , p.threshold = 1 )
plot_heatmap_cell( res = filter.binary$binary.res , max.group = "O" , merge = F , trend = 'Down' , p.threshold = 1 )

plot_heatmap_cell( res = filter.anova$anova.res , max.group = "O" , merge = T , p.threshold = 1 )

}
