% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_line}
\alias{plot_line}
\title{Line plot}
\usage{
plot_line(
  res,
  CCC.ID,
  p = T,
  p.adj = T,
  colors = "blue",
  dot.size = 2,
  model.text = 3,
  axis.text = 8,
  axis.title = 10,
  plot.title = 12
)
}
\arguments{
\item{res}{The object returned by the filterCCC function.}

\item{CCC.ID}{The ID of the CCC (cell–cell communication) event.}

\item{p}{Whether to display the p-value.}

\item{p.adj}{Whether to display the p-adjusted.}

\item{colors}{The color of the line.}

\item{dot.size}{The size of the dots.}

\item{model.text}{The font size of the model results.}

\item{axis.text}{The font size of the axis labels.}

\item{axis.title}{The font size of the axis titles.}

\item{plot.title}{The font size of the plot title.}
}
\value{
A ggplot2 object.
}
\description{
Plot a fitted curve for the continuous variable.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
ccc.glm <- multiCCC( data = LRscore , glm.column = 'weight' )
ccc.time <- multiCCC( data = LRscore , time.course.params = list( time  = 'time' , replicate  = 'replicate'  ) )

#3.filterCCC
filter.glm <- filterCCC(ccc.glm )
filter.time <- filterCCC(ccc.time)

#4.plot_line
CCC.ID <- 'cell2_cell3.CALR_ITGAV'
plot_line( res = filter.glm$glm.res,
           CCC.ID = CCC.ID , p.adj = F )

}
