% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.network.R
\name{plot_network}
\alias{plot_network}
\title{Network plot of sender-receiver}
\usage{
plot_network(
  res,
  data.used = "filtered.result",
  padj.threshold = 0.05,
  p.threshold = NULL,
  max.group = NULL,
  ligand = NULL,
  receptor = NULL,
  sender = NULL,
  receiver = NULL,
  CCC.ID = NULL,
  title = NULL,
  plot.title = 14,
  merge = T,
  link.color = c("#FF0000", "#085AFF"),
  link.alpha = 0.8,
  link.weight.range = c(1, 5),
  node.size = NULL,
  node.border.size = 1,
  node.border.color = "grey",
  node.fill = "white",
  label.size = 5,
  label.color = "black"
)
}
\arguments{
\item{res}{The object returned by the filterCCC function.}

\item{data.used}{The data used for plotting. Options are result or filtered.result.}

\item{padj.threshold}{The threshold for p.adj.}

\item{p.threshold}{Similar to padj.threshold, this parameter specifies the p-value threshold. When provided, padj.threshold will be ignored.}

\item{max.group}{Which group is treated as the communication-enhanced group.}

\item{ligand}{Character vector for filtering CCC.}

\item{receptor}{Character vector for filtering CCC.}

\item{sender}{Character vector for filtering CCC.}

\item{receiver}{Character vector for filtering CCC.}

\item{CCC.ID}{Character vector for filtering CCC.}

\item{title}{Plot title.}

\item{plot.title}{The font size of the plot title.}

\item{merge}{Whether upregulated and downregulated CCC events are displayed in a single plot.}

\item{link.color}{Provide a two-value color vector to specify the fill colors of links.}

\item{link.alpha}{A numeric value specifying the transparency of the links.}

\item{link.weight.range}{Provide a two-value numeric vector specifying the range of link widths.}

\item{node.size}{Node size.}

\item{node.border.size}{Width of the node outline.}

\item{node.border.color}{Color of the node outline.}

\item{node.fill}{Fill color of the node.}

\item{label.size}{Size of the text inside the node.}

\item{label.color}{Color of the text inside the node.}
}
\value{
A ggplot2 object.
}
\description{
A network visualization of cell–cell interactions for categorical variables.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )

#3.filterCCC
filter.binary <- filterCCC(ccc.binary)
filter.anova <- filterCCC(ccc.anova)

#4.network
plot_network( res = filter.binary$binary.res , p.threshold = 1  )
plot_network( res = filter.binary$binary.res , p.threshold = 1 , merge = F )


}
