% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_score.R
\name{scoreLR}
\alias{scoreLR}
\title{Compute LRscore}
\usage{
scoreLR(
  exp,
  meta.data,
  sample,
  celltype,
  LR.species = "human",
  LR.source = "Consensus",
  LR.method = "SingleCellSignalR",
  min.cell = 10,
  min.exp = 0.1,
  min.prob = 0.3,
  threads = NULL
)
}
\arguments{
\item{exp}{A matrix object with cells as row names and genes as column names.}

\item{meta.data}{A data.frame object whose row names are identical to those of the exp parameter.}

\item{sample}{The column name in meta.data that represents the sample ID.}

\item{celltype}{The column name in meta.data that represents the cell type.}

\item{LR.species}{Species. Options are human or mouse.}

\item{LR.source}{Two options:

(1) Ligand–receptor resource (with CCI or Consensus recommended). Available options include: CCI, Consensus, Baccin2019, CellCall, CellChatDB, Cellinker, CellPhoneDB, CellTalkDB, connectomeDB2020, EMBRACE, Guide2Pharma, HPMR, ICELLNET, iTALK, Kirouac2010, LRdb, Ramilowski2015, OmniPath, and MouseConsensus. For mouse, only the CCI and MouseConsensus options are available.

(2) A data.frame object can be provided, in which the first column contains ligand gene names and the second column contains receptor gene names.}

\item{LR.method}{Method for calculating LR score. Available options include: CCI, SingleCellSignalR, connectome, iTALK, NATMI, CytoTalk, cellchat, and cellphoneDB. Note that cellchat and cellPhoneDB are relatively time-consuming to run.}

\item{min.cell}{The minimum number of cells expressing the ligand or receptor.}

\item{min.exp}{The minimum expression level of ligands and receptors included in the analysis.}

\item{min.prob}{The minimum proportion of cells expressing the ligand and receptor. Note: At least one of min.cell, min.exp, or min.prob must be non-zero.}

\item{threads}{Number of cores used for parallel computation. By default, the maximum computing resources are used.}
}
\value{
A list object.
}
\description{
Compute cell–cell communication score (LRscore).
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

}
