% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCCC.R
\name{multiCCC}
\alias{multiCCC}
\title{Detect differential CCC events}
\usage{
multiCCC(
  data,
  binary.params = NULL,
  anova.column = NULL,
  glm.column = NULL,
  time.course.params = NULL,
  covariance = NULL,
  p.adjust.method = "BH",
  threads = NULL
)
}
\arguments{
\item{data}{The object returned by the scoreLR function.}

\item{binary.params}{A list defining the comparison. It should contain at least \code{group}, \code{g1}, and \code{g2}, where \code{group} specifies a column in \code{meta.data}, and the fold change (FC) is calculated as \code{g1 / g2}.
The optional element \code{test} specifies the statistical test:

\itemize{
\item Wilcoxon rank-sum test (default): list(group = '', g1 = '', g2 = '', test = 'wilcoxon' )
\item Student's t-test: list(group = '', g1 = '', g2 = '', test = 't' )
\item Permutation test: list(group = '', g1 = '', g2 = '', test = 'permutation' )
}}

\item{anova.column}{A column name in meta.data based on which an ANOVA test is performed.}

\item{glm.column}{A column name in meta.data used to fit a GLM model.}

\item{time.course.params}{Provide a list object with the format list(time = '', replicate = ''), where both time and replicate are column names in meta.data of numeric type. time represents the time point, and replicate represents the biological replicate.}

\item{covariance}{Column names in meta.data representing covariates, used only for GLM modeling and time course analysis.}

\item{p.adjust.method}{The method for multiple testing correction, as detailed in the p.adjust() function.}

\item{threads}{Number of cores used for parallel computation. By default, the maximum computing resources are used.}
}
\value{
A list object.

Note: The multiCCC function supports the simultaneous use of the binary.params, anova.column, glm.column, and time.course.params parameters, and analyses will be performed automatically for each application scenario.
}
\description{
Detect differential CCC (cell–cell communication) events.
}
\examples{

data( "CCC.test.data" )

#1.scoreLR
LRscore <- scoreLR( exp = t( CCC.test.data$exp ) ,
                 meta.data = CCC.test.data$meta ,
                 LR.species = 'human' ,
                 sample = 'orig.ident' ,
                 celltype = 'celltype' )

#2.multiCCC
#Note: The multiCCC function supports the simultaneous use of the binary.params, anova.column, glm.column, and time.course.params parameters, and analyses will be performed automatically for each application scenario.
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )
ccc.glm <- multiCCC( data = LRscore , glm.column = 'weight' )
ccc.time <- multiCCC( data = LRscore , time.course.params = list( time  = 'time' , replicate  = 'replicate'  ) )

#One-step execution for all scenarios
ccc.all <- multiCCC( data = LRscore , 
	   binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ),
	   anova.column = 'batch',
	   glm.column = 'weight',
	   time.course.params = list( time  = 'time' , replicate  = 'replicate'  )
	)

}
