% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCCC.R
\name{multiCCC}
\alias{multiCCC}
\title{Detect differential CCC (cell–cell communication) events}
\usage{
multiCCC(
  data,
  binary.params = NULL,
  anova.column = NULL,
  glm.column = NULL,
  time.course.params = NULL,
  covariance = NULL,
  p.adjust.method = "BH",
  threads = NULL
)
}
\arguments{
\item{data}{The object returned by the lr_score function.}

\item{binary.params}{A list object with the format list(group = '', g1 = '', g2 = ''), where group represents a column name in meta.data, and the fold change (FC) in the results is calculated as g1/g2.}

\item{anova.column}{A column name in meta.data based on which an ANOVA test is performed.}

\item{glm.column}{A column name in meta.data used to fit a GLM model.}

\item{time.course.params}{Provide a list object with the format list(time = '', replicate = ''), where both time and replicate are column names in meta.data of numeric type. time represents the time point, and replicate represents the biological replicate.}

\item{covariance}{Column names in meta.data representing covariates, used only for GLM modeling and time course analysis.}

\item{p.adjust.method}{The method for multiple testing correction, as detailed in the p.adjust() function.}

\item{threads}{The number of threads used for parallel computation.}
}
\value{
A list object.
}
\description{
Detect differential CCC (cell–cell communication) events
}
\examples{
ccc.binary <- multiCCC( data = LRscore , binary.params = list( group = 'Group' , g1 = 'O' , g2 = 'Y'  ) )
ccc.anova <- multiCCC( data = LRscore , anova.column = 'batch' )
ccc.glm <- multiCCC( data = LRscore , glm.column = 'weight' )
ccc.time <- multiCCC( data = data , time.course.params = list( time  = 'time' , replicate  = 'replicate'  ) )


}
