% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dot}
\alias{plot_dot}
\title{Dot plot}
\usage{
plot_dot(
  ccc.res,
  fill = "p.adj",
  threshold = 0.05,
  colors = c("#0000FF", "#F7F7F7", "#FF0000"),
  ligand = NULL,
  receptor = NULL,
  source.cell = NULL,
  target.cell = NULL,
  CCC.ID = NULL,
  dot.max = 3,
  dot.min = 1,
  legend.position = "right",
  strip.text = 10,
  x.text = 10,
  y.text = 10,
  facet.row = 1
)
}
\arguments{
\item{ccc.res}{The object returned by the multiCCC function.}

\item{fill}{The column used for color fill, optionally p or p.adj.}

\item{threshold}{The threshold for p or p.adj.}

\item{colors}{A character vector.}

\item{ligand}{A character vector.}

\item{receptor}{A character vector.}

\item{source.cell}{A character vector.}

\item{target.cell}{A character vector.}

\item{CCC.ID}{A character vector.}

\item{dot.max}{The maximum value of the dots.}

\item{dot.min}{The minimum value of the dots.}

\item{legend.position}{The position of the legend in the plot.}

\item{strip.text}{The font size of the facet labels.}

\item{x.text}{The font size of the X-axis labels.}

\item{y.text}{The font size of the Y-axis labels.}

\item{facet.row}{The number of rows in the facets.}
}
\value{
A ggplot2 object.
}
\description{
Plot a dot plot for the results of a categorical variable.
}
\examples{
plot_dot( ccc.res = ccc.anova$anova.res , ligand = c('Sele' , 'Fgf1' , 'Nrxn1' ,'Nrxn2' ), threshold =  1 ,strip.text = 8 )


}
