% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_line}
\alias{plot_line}
\title{Plot a fitted curve for a continuous variable.}
\usage{
plot_line(
  ccc.res,
  CCC.ID,
  p = T,
  p.adj = T,
  colors = "blue",
  dot.size = 2,
  model.text = 3,
  axis.text = 8,
  axis.title = 10,
  plot.title = 12
)
}
\arguments{
\item{ccc.res}{The object returned by the multiCCC function.}

\item{CCC.ID}{The ID of the CCC (cell–cell communication) event.}

\item{p}{Whether to display the p-value.}

\item{p.adj}{Whether to display the p-adjusted.}

\item{colors}{The color of the line.}

\item{dot.size}{The size of the dots.}

\item{model.text}{The font size of the model results.}

\item{axis.text}{The font size of the axis labels.}

\item{axis.title}{The font size of the axis titles.}

\item{plot.title}{The font size of the plot title.}
}
\value{
A ggplot2 object.
}
\description{
Plot a fitted curve for a continuous variable.
}
\examples{
CCC.ID <- 'Endothelial cell_Monocyte.Nid1_Itgav'
plot_line( ccc.res = cci.glm$glm.res, CCC.ID = CCC.ID , p.adj = F )

}
